Improve error handling and fix robot exit bug
- Moved error messages to ErrorMessages class
- Added global exception handler
- Added NativeException for library loading exception
-

+++ handsfree/src/main/java/ch/bbcag/handsfree/jni/NativeUtils.java
@@ -1,5 +1,7 @@
 package ch.bbcag.handsfree.jni;
 
+import ch.bbcag.handsfree.error.NativeException;
+
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -7,28 +9,36 @@ import java.io.InputStream;
 
 public class NativeUtils {
 
-    public static void loadLibraryFromResource(String resource) throws IOException {
+    public static void loadLibraryFromResource(String resource) throws NativeException {
         InputStream in = NativeUtils.class.getResourceAsStream(resource);
-        byte[] buffer = new byte[in.available()];
-        in.read(buffer);
-        in.close();
-
-        String resourceName = resource.substring(resource.lastIndexOf("/") + 1);
-        String prefix = resourceName.split("\\.")[0];
-        String suffix = "." + resourceName.split("\\.")[1];
-
-        File temporaryFile = File.createTempFile(prefix, suffix);
-        temporaryFile.deleteOnExit();
-        if(!temporaryFile.exists()) {
-            temporaryFile.getParentFile().mkdirs();
-            temporaryFile.createNewFile();
+        if(in == null) {
+            throw new NativeException(resource, "Couldn't find resource: " + resource);
         }
-        temporaryFile.deleteOnExit();
-        FileOutputStream out = new FileOutputStream(temporaryFile);
-        out.write(buffer);
-        out.close();
 
-        System.load(temporaryFile.getAbsolutePath());
+        try {
+            byte[] buffer = new byte[in.available()];
+            in.read(buffer);
+            in.close();
+
+            String resourceName = resource.substring(resource.lastIndexOf("/") + 1);
+            String prefix = resourceName.split("\\.")[0];
+            String suffix = "." + resourceName.split("\\.")[1];
+
+            File temporaryFile = File.createTempFile(prefix, suffix);
+            temporaryFile.deleteOnExit();
+            if(!temporaryFile.exists()) {
+                temporaryFile.getParentFile().mkdirs();
+                temporaryFile.createNewFile();
+            }
+            temporaryFile.deleteOnExit();
+            FileOutputStream out = new FileOutputStream(temporaryFile);
+            out.write(buffer);
+            out.close();
+
+            System.load(temporaryFile.getAbsolutePath());
+        } catch(IOException e) {
+            throw new NativeException(resource, e.getMessage());
+        }
     }
 
 }
